removed helper service and added inside base command
