+++ package.json
@@ -51,7 +51,7 @@
                 "category": "debricked"
             },
             {
-                "command": "debricked.help",
+                "command": "debricked.debricked.help",
                 "title": "Help about any command",
                 "category": "debricked"
             },
