+++ src/services/index.ts
@@ -1,6 +1,5 @@
 import { BaseCommandService } from "./baseCommandService";
 import { FileService } from "./fileService";
-import { HelpService } from "./helpService";
 import { ScanService } from "./scanService";
 
-export { BaseCommandService, HelpService, ScanService, FileService };
+export { BaseCommandService, ScanService, FileService };
